+++ Source/TitaniumKit/src/UI/ViewLayoutDelegate.cpp
@@ -300,12 +300,12 @@ namespace Titanium
 			autoLayoutForHeight__ = value;
 		}
 
-		double ViewLayoutDelegate::get_zIndex() const TITANIUM_NOEXCEPT
+		int32_t ViewLayoutDelegate::get_zIndex() const TITANIUM_NOEXCEPT
 		{
 			return zIndex__;
 		}
 
-		void ViewLayoutDelegate::set_zIndex(const double& zIndex) TITANIUM_NOEXCEPT
+		void ViewLayoutDelegate::set_zIndex(const int32_t& zIndex) TITANIUM_NOEXCEPT
 		{
 			zIndex__ = zIndex;
 		}
